feat: add admin fields

+++ src/courses/ui/edit-config/components/edit-couse-config-dialog.tsx
@@ -24,9 +24,9 @@ import { editCourseConfigAction } from "../actions/edit-course-config-action";
 
 const EditCourseConfigSchema = z.object({
   enableFuzz: z.boolean(),
-  maximumInterval: z.number().min(1).max(default_maximum_interval),
+  maximumInterval: z.number().int().min(1).max(default_maximum_interval),
   requestRetention: z.number().min(0).max(1),
-  dailyNewCardsCount: z.number().min(1).max(100),
+  dailyNewCardsCount: z.number().int().min(1).max(100),
   showAdvancedRatingOptions: z.boolean(),
 });
