+++ src/notes/ui/course-notes/schemas/paginate-notes-action-model.ts
@@ -6,8 +6,8 @@ import { ObjectIdSchema } from "@/src/common/schemas/object-id-schema";
  */
 export const PaginateNotesActionSchema = z.object({
   courseId: ObjectIdSchema,
-  page: z.number().optional(),
-  pageSize: z.number().optional(),
+  page: z.number().int().optional(),
+  pageSize: z.number().int().optional(),
 });
 
 /**
